+++ core/src/com/meowstgdx/game/inventory/items/weapons/MonkeyFist.java
@@ -1,0 +1,36 @@
+package com.meowstgdx.game.inventory.items.weapons;
+
+import com.meowstgdx.game.materials.ToolMaterial;
+
+public class MonkeyFist extends Weapon {
+    public MonkeyFist(ToolMaterial toolMaterial) {
+        super("MonkeyFist", 16, 0.25f, 64, toolMaterial);
+
+        switch (toolMaterial) {
+            case WOOD: {
+                this.damage *= 1;
+                this.durability *= 1;
+            }
+            case FLINT: {
+                this.damage *= 2;
+                this.durability *= 2;
+            }
+            case STONE: {
+                this.damage *= 3;
+                this.durability *= 3;
+            }
+            case BRASS: {
+                this.damage *= 4;
+                this.durability *= 4;
+            }
+            case IRON: {
+                this.damage *= 5;
+                this.durability *= 5;
+            }
+            case TITANIUM: {
+                this.damage *= 10;
+                this.durability *= 10;
+            }
+        }
+    }
+}