+++ core/src/com/meowstgdx/game/items/weapons/Staff.java
@@ -1,4 +1,0 @@
-package com.meowstgdx.game.items.weapons;
-
-public class Staff {
-}