+++ changelog.md
@@ -6,7 +6,8 @@
 
 **Improvements**
 -
-- Mob texture states are improved on the backend, using "facingRight" and "holdingItem" states to change textures accordingly
+- Hotbar now has visible slots that can be accessed by the player, and can contain items that the player has
+- Player is now able to visibly hold items when selected in the hotbar
 
 ***